# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ops & Planning API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/setup:
    post:
      operationId: setupFirstAdmin
      tags: [auth]
      summary: Create the first admin account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetupBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "503": { $ref: "#/components/responses/Error" }
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Email + password login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Invalidate current session
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: No content }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Current authenticated user
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/forgot-password:
    post:
      operationId: requestPasswordReset
      tags: [auth]
      summary: Request a password-reset link
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ForgotPasswordBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Complete a password reset
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResetPasswordBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /auth/accept-invite:
    post:
      operationId: acceptInvite
      tags: [auth]
      summary: Activate an invited account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AcceptInviteBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
  /auth/invite:
    post:
      operationId: createInvite
      tags: [auth, invites]
      summary: Admin creates an invite
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateInviteBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreateInviteResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
  /auth/invite/{token}:
    get:
      operationId: getInviteByToken
      tags: [auth, invites]
      summary: Look up an invite by its token
      parameters:
        - { in: path, name: token, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitePreview" }
        "400": { $ref: "#/components/responses/BadRequest" }

  # ─── Users ────────────────────────────────────────────────────────────
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/EventWithVisibility" }
    post:
      operationId: createEvent
      tags: [events]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateEventBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventWithVisibility" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateEvent
      tags: [events]
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateEventBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteEvent
      tags: [events]
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }

  # ─── Team Notes ───────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Conflict:
      description: Conflict
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }
    UserRole:
      type: string
      enum: [admin, stream_overseer, leader]
...
    AuthSession:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user: { $ref: "#/components/schemas/User" }
    User:
      type: object
      required: [id, email, name, initials, department, role, active, createdAt, updatedAt]
      properties:
        id: { type: string }
        email: { type: string, format: email }
        name: { type: string }
        initials: { type: string }
        department: { type: string }
        role: { $ref: "#/components/schemas/UserRole" }
        streamId: { type: string, nullable: true }
        teamId: { type: string, nullable: true }
        active: { type: boolean }
        invitedByName: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    LoginBody:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }
    SetupBody:
      type: object
      required: [email, name, password, setupSecret]
      properties:
        email: { type: string, format: email }
        name: { type: string, minLength: 1 }
        password: { type: string, minLength: 8 }
        department: { type: string }
        setupSecret: { type: string, minLength: 1 }
    AcceptInviteBody:
      type: object
      required: [token, name, password]
      properties:
        token: { type: string, minLength: 1 }
        name: { type: string, minLength: 1 }
        password: { type: string, minLength: 8 }
    ForgotPasswordBody:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    ResetPasswordBody:
      type: object
      required: [token, password]
      properties:
        token: { type: string, minLength: 1 }
        password: { type: string, minLength: 8 }
    CreateInviteBody:
      type: object
      required: [email, name, role]
      properties:
        email: { type: string, format: email }
        name: { type: string, minLength: 1 }
        role: { $ref: "#/components/schemas/UserRole" }
        department: { type: string }
        streamId: { type: string, nullable: true }
        teamId: { type: string, nullable: true }
    CreateInviteResponse:
      type: object
      required: [message, acceptUrl]
      properties:
        message: { type: string }
        acceptUrl: { type: string }
    InvitePreview:
      type: object
      required: [email, name, role, department]
      properties:
        email: { type: string, format: email }
        name: { type: string }
        role: { $ref: "#/components/schemas/UserRole" }
        department: { type: string }
...
    EventStatus:
      type: string
      enum: [pending, approved, rejected]
    EventVisibility:
      type: string
      enum: [full, shared]
    Event:
      type: object
      required:
        - id
        - title
        - sharedDescription
        - startDate
        - endDate
        - location
        - color
        - isAllDay
        - status
        - invitedTeamIds
        - createdAt
        - updatedAt
      properties:
        id: { type: string }
        title: { type: string }
        internalDescription: { type: string, nullable: true }
        sharedDescription: { type: string }
        startDate: { type: string, format: date-time }
        endDate: { type: string, format: date-time }
        location: { type: string }
        color: { type: string }
        isAllDay: { type: boolean }
        status: { $ref: "#/components/schemas/EventStatus" }
        projectId: { type: string, nullable: true }
        createdByTeamId: { type: string, nullable: true }
        createdByUserId: { type: string, nullable: true }
        invitedTeamIds:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    EventWithVisibility:
      allOf:
        - $ref: "#/components/schemas/Event"
        - type: object
          required: [visibility]
          properties:
            visibility: { $ref: "#/components/schemas/EventVisibility" }
    CreateEventBody:
      type: object
      required: [title, startDate, endDate]
      properties:
        title: { type: string, minLength: 1 }
        internalDescription: { type: string }
        sharedDescription: { type: string }
        startDate: { type: string, format: date-time }
        endDate: { type: string, format: date-time }
        location: { type: string }
        color: { type: string }
        isAllDay: { type: boolean }
        status: { $ref: "#/components/schemas/EventStatus" }
        projectId: { type: string, nullable: true }
        invitedTeamIds:
          type: array
          items: { type: string }
    UpdateEventBody:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        internalDescription: { type: string }
        sharedDescription: { type: string }
        startDate: { type: string, format: date-time }
        endDate: { type: string, format: date-time }
        location: { type: string }
        color: { type: string }
        isAllDay: { type: boolean }
        status: { $ref: "#/components/schemas/EventStatus" }
        projectId: { type: string, nullable: true }
        invitedTeamIds:
          type: array
          items: { type: string }
...