# lib/api-zod/src/generated/types/authSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ops & Planning API specification
 * OpenAPI spec version: 0.1.0
 */
import type { User } from "./user";

export interface AuthSession {
  token: string;
  user: User;
}