# lib/api-zod/src/generated/types/resetPasswordBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ops & Planning API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ResetPasswordBody {
  /** @minLength 1 */
  token: string;
  /** @minLength 8 */
  password: string;
}