# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ops & Planning API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: string;
  email: string;
  name: string;
  initials: string;
  department: string;
  role: UserRole;
  streamId?: string | null;
  teamId?: string | null;
  active: boolean;
  invitedByName?: string | null;
  createdAt: Date;
  updatedAt: Date;
}